+++ examples/memory/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## [0.10.31](https://github.com/AIGNE-io/aigne-framework/compare/example-memory-v0.10.30...example-memory-v0.10.31) (2025-08-20)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.23
+    * @aigne/cli bumped to 1.37.1
+    * @aigne/core bumped to 1.52.0
+    * @aigne/default-memory bumped to 1.1.5
+    * @aigne/fs-memory bumped to 1.0.23
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.31
+
+## [0.10.30](https://github.com/AIGNE-io/aigne-framework/compare/example-memory-v0.10.29...example-memory-v0.10.30) (2025-08-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @aigne/agent-library bumped to 1.21.22
+    * @aigne/cli bumped to 1.37.0
+    * @aigne/core bumped to 1.51.0
+    * @aigne/default-memory bumped to 1.1.4
+    * @aigne/fs-memory bumped to 1.0.22
+  * devDependencies
+    * @aigne/test-utils bumped to 0.5.30
+
 ## [0.10.29](https://github.com/AIGNE-io/aigne-framework/compare/example-memory-v0.10.28...example-memory-v0.10.29) (2025-08-16)